+++ scripts/bin/chef/_delete.sh
@@ -5,43 +5,61 @@
 #  Specification Path      : N/A_____________________________________________________________________|
 #  Author                  : brady house_____________________________________________________________|
 #  Create date             : 11/21/2018______________________________________________________________|
-#  Description             : MASTER CHEF TEST FUNCTION_______________________________________________|
-#  Entry Point             : reactStart______________________________________________________________|
+#  Description             : MASTER CHEF DELETE FUNCTION_____________________________________________|
+#  Entry Point             : destroy_________________________________________________________________|
 #  Input Parameters        : N/A_____________________________________________________________________|
-#  Initial Consumer        : ../fiddle-test.sh_______________________________________________________|
+#  Initial Consumer        : ../fiddle-delete.sh_____________________________________________________|
 # ---------------------------------------------------------------------------------------------------|
 #  Revision History::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::|
 # ---------------------------------------------------------------------------------------------------|
 # Baseline Ver - See CHANGELOG @ 262_add_chef_setup
 # ---------------------------------------------------------------------------------------------------|
 
-function nvmInstall() {
-  groupLog "nvmInstall";
+function deleteFiddle() {
+   groupLog "deleteFiddle"
+   _curDir=$(pwd;);
+   if [[ ! $(which kitchen;) ]]
+   then
+        rm -rf "$1" || exit $?;
+   else
+        cd "$1" || exit $?;
+        kitchen destroy || exit $?;
+        cd ${_curDir} || exit $?;
+        rm -rf "$1" || exit $?;
+   fi
+}
 
-  if [[ -d ${NVM_DIR} ]]
+
+function catch() {
+    case $1 in
+        0)  endLog "\"${fiddleSubDir}\" delete.";
+            ;;
+        *)  endLog "fubar! Something went wrong.";
+            ;;
+    esac
+    exit $1
+}
+
+function destroy() {
+  if [[ "$#" -ne 1 ]]
   then
-    source ${NVM_DIR}/nvm.sh;
-    nvm install ${NVM_VERSION};
-  else
-    exit 3;
+        echo "Incorrect number of arguments"
+        echo "Please specify the name of the fiddle to be deleted"
+        exit 59
   fi
+
+  # try
+  (
+      deleteFiddle $1 || exit $?;
+  )
+  # catch
+  rc=$?; catch ${rc};
+
+  #finally
+  exit ${rc};
+
+
+
 }
 
 
-
-
-function npmInstall() {
-  groupLog "npmInstall";
-  if [[ ! -d node_modules ]]
-  then
-    npm install;
-  fi
-}
-
-function reactTest() {
-    groupLog "reactTest";
-    nvmInstall || exit $?;
-    npmInstall || exit $?;
-    npm test || exit $?;
-    exit 0;
-}